+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+### Added
+- Handle `Symfony\Component\HttpKernel\Exception\HttpException` exceptions.
+
 ## [5.1.0] - 2021-03-29
 
 ### Added